backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
Pillow